Rename tooltip.ts -> lock.ts
