+++ apps/desktop/src/lib/files/lock.ts
